# parser: batch some allocations in the scanner