The scanner allocates copies of all uppercase identifiers; preallocate
a buffer (proportional to the size of the query) to save some of these
allocations.

Benchmark:
```
name   old time/op    new time/op    delta
Parse    33.8µs ± 1%    33.7µs ± 1%     ~     (p=0.699 n=6+6)

name   old alloc/op   new alloc/op   delta
Parse    5.54kB ± 0%    5.68kB ± 0%   +2.60%  (p=0.002 n=6+6)

name   old allocs/op  new allocs/op  delta
Parse      92.0 ± 0%      81.0 ± 0%  -11.96%  (p=0.002 n=6+6)
```

Release note: None