@RaduBerinde 2018-12-04
___
*[pkg/sql/parser/scan.go, line 826 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/32805#-LSqabdN3cblaR8Zpp9t:-LSuVW2d5PmGKo1esRUg:bitqcvi) ([raw file](https://github.com/cockroachdb/cockroach/blob/d45ab86b886b2c35eb3c65bb96b09a515032ed96/pkg/sql/parser/scan.go#L826)):*
<details><summary><i>Previously, nvanbenschoten (Nathan VanBenschoten) wrote…</i></summary><blockquote>

> Or are you suggesting using a preallocated slice and then making a copy of the right size at the end?

Do we even need the copy? Would it be ok to perform an unsafe bytes -> string cast?
</blockquote></details>

This is in the context of your suggestion of reusing a slice. We can reuse, or we can do an unsafe cast, but not both..


<!-- Sent from Reviewable.io -->